Refactor to package by feature
