+++ app/src/main/java/com/mathewsmobile/pwned/constants/Constants.kt
@@ -1,4 +1,4 @@
-package com.mathewsmobile.pwned.util
+package com.mathewsmobile.pwned.constants
 
 import android.graphics.Color
