+++ app/src/main/java/com/mathewsmobile/pwned/shared/SingleFragmentActivity.kt
@@ -1,4 +1,4 @@
-package com.mathewsmobile.pwned.activities
+package com.mathewsmobile.pwned.shared
 
 import android.os.Bundle
 import android.os.PersistableBundle
